pyproject.toml
[project]
name = "ai-text-buddy-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi[standard]>=0.116.0",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "pydantic>=2.11.7",
]

[project.optional-dependencies]
lambda = [
    "mangum>=0.17.0"
]
dev = [
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0"
]
